ocr/requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.36.2
pillow==10.0.1
opencv-python==4.8.1.78
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pdf2image==1.16.3
poppler-utils
surya-ocr
pytesseract